New model not yet ready for deployment
Compared to currently deployed model, still seeing too much incorrect assignment to the Company IM and Extremely Casual categories. I'm fairly sure that the next logical step is to combine those two categories or drop one from the model. As this would require a lot of reconfiguring in the app code, I will need to save that thought for another day.